+++ api/src/controllers/EspecController/especialidadController.js
@@ -1,4 +1,4 @@
-const { Especialidad } = require("../db");
+const { Especialidad } = require("../../db");
 
 async function getEspecialidad(req, res) {
   try {
