+++ api/src/controllers/TipoController/deleteTipoController.js
@@ -1,4 +1,4 @@
-const { Tipo } = require("../db");
+const { Tipo } = require("../../db");
 
 const deleteTipo = async (req, res) => {
   const { id } = req.params;
