+++ api/src/controllers/index.js
@@ -1,19 +1,19 @@
-const getMenu = require("./menuController");
-const getEspecialidad = require("./especialidadController");
-const getTipo = require("./tipoController");
-const postEspecialidad = require("./postEspecialidad");
-const postTipo = require("./postTipo");
-const findMenuById = require("./findMenuByIdController");
-const createMenu = require("./createMenuController");
-const getMenusByQuery = require("./getMenusByQuery");
-const updateTipo = require("./updateTipoController");
-const deleteTipo = require("./deleteTipoController");
-const updateEspecialidad = require("./updateEspecialidad");
-const deleteEspecialidad = require("./deleteEspecialidad");
-const registerCtrl = require("./registerController");
-const loginCtrl = require("./loginController");
-const postRole = require("./postRoleController");
-const getUserByEmail = require("./getUserByEmail");
+const getMenu = require("./MenuController/menuController");
+const getEspecialidad = require("./EspecController/especialidadController");
+const getTipo = require("./TipoController/tipoController");
+const postEspecialidad = require("./EspecController/postEspecialidad");
+const postTipo = require("./TipoController/postTipo");
+const findMenuById = require("./MenuController/findMenuByIdController");
+const createMenu = require("./MenuController/createMenuController");
+const getMenusByQuery = require("./MenuController/getMenusByQuery");
+const updateTipo = require("./TipoController/updateTipoController");
+const deleteTipo = require("./TipoController/deleteTipoController");
+const updateEspecialidad = require("./EspecController/updateEspecialidad");
+const deleteEspecialidad = require("./EspecController/deleteEspecialidad");
+const registerCtrl = require("./AuthController/registerController");
+const loginCtrl = require("./AuthController/loginController");
+const postRole = require("./AuthController/postRoleController");
+const getUserByEmail = require("./AuthController/getUserByEmail");
 module.exports = {
   getMenu,
   findMenuById,
